# src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Learning Management System Authentication APIs",
		"description": "This is a Learning Management System API documentation which is only exposed on development environment.",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local host server, exclusively for back-end developers."
		},
		{
			"url": "https://smooth-raptor-explicitly.ngrok-free.app/api/auth",
			"description": "LMS API's related to user authentications"
		}
	],
	"paths": {
		"/student/register": {
			"post": {
				"summary": "Register a student",
				"tags": [
					"Student"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"example": {
									"firstName": "John",
									"lastName": "Doe",
									"email": "johndoe@gmail.com",
									"password": "john@123"
								},
								"required": [
									"firstName",
									"lastName",
									"email",
									"password"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"example": "3a8a46f8-b4c3-4ff7-b900-99b9ee07efe8"
												},
												"firstName": {
													"type": "string",
													"example": "John"
												},
												"lastName": {
													"type": "string",
													"example": "Doe"
												},
												"email": {
													"type": "string",
													"example": "johndoe@gmail.com"
												},
												"accessToken": {
													"type": "string",
													"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
												},
												"refreshToken": {
													"type": "string",
													"example": "9fe94329-0fb3-4391-96a8-be05b9be5a5b"
												}
											},
											"required": [
												"id",
												"firstName",
												"lastName",
												"email",
												"accessToken",
												"refreshToken"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/student/sign-in": {
			"post": {
				"summary": "Sign-in a student with email",
				"tags": [
					"Student"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"example": {
									"email": "johndoe@gmail.com",
									"password": "john@123"
								},
								"required": [
									"email",
									"password"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"example": "3a8a46f8-b4c3-4ff7-b900-99b9ee07efe8"
												},
												"firstName": {
													"type": "string",
													"example": "John"
												},
												"lastName": {
													"type": "string",
													"example": "Doe"
												},
												"email": {
													"type": "string",
													"example": "johndoe@gmail.com"
												},
												"accessToken": {
													"type": "string",
													"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
												},
												"refreshToken": {
													"type": "string",
													"example": "9fe94329-0fb3-4391-96a8-be05b9be5a5b"
												}
											},
											"required": [
												"id",
												"firstName",
												"lastName",
												"email",
												"accessToken",
												"refreshToken"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/student/sign-in/gmail": {
			"post": {
				"summary": "Sign-in a student with gmail",
				"tags": [
					"Student"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"authCode": {
										"type": "string"
									},
									"redirectUri": {
										"type": "string"
									}
								},
								"example": {
									"authCode": "4/0AY0e-g4nW8gRmX2zQ5VwHPLp3yDt_mwZ6yZ-GfbLwdrgW_3v6jZVJf43KLmEr39VTuUoTwFAUsnExFwVeZu_5wJpM",
									"redirectUri": "http://localhost:3000"
								},
								"required": [
									"authCode",
									"redirectUri"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"example": "3a8a46f8-b4c3-4ff7-b900-99b9ee07efe8"
												},
												"firstName": {
													"type": "string",
													"example": "John"
												},
												"lastName": {
													"type": "string",
													"example": "Doe"
												},
												"email": {
													"type": "string",
													"example": "johndoe@gmail.com"
												},
												"accessToken": {
													"type": "string",
													"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
												},
												"refreshToken": {
													"type": "string",
													"example": "9fe94329-0fb3-4391-96a8-be05b9be5a5b"
												}
											},
											"required": [
												"id",
												"firstName",
												"lastName",
												"email",
												"accessToken",
												"refreshToken"
											]
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/student/forgot-password": {
			"post": {
				"summary": "Send forgot password email for a student",
				"tags": [
					"Student"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"example": {
									"email": "johndoe@gmail.com"
								},
								"required": [
									"email"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"nullable": true
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/student/reset-password": {
			"post": {
				"summary": "Reset Password for a student",
				"tags": [
					"Student"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"verificationCode": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"example": {
									"email": "johndoe@gmail.com",
									"verificationCode": "MK67",
									"password": "john123"
								},
								"required": [
									"email",
									"verificationCode",
									"password"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"nullable": true
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		},
		"/student/": {
			"get": {
				"summary": "Retrieve student profile",
				"tags": [
					"Student"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"$ref": "#/components/schemas/User"
										}
									},
									"required": [
										"data"
									]
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"requestBodies": {},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"errors": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"code": {
									"type": "string"
								},
								"message": {
									"type": "string"
								},
								"field": {
									"type": "string"
								}
							},
							"required": [
								"code",
								"message"
							]
						}
					}
				},
				"required": [
					"errors"
				],
				"example": {
					"errors": [
						{
							"code": "INV_INP",
							"message": "Name is required"
						}
					]
				}
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "3a8a46f8-b4c3-4ff7-b900-99b9ee07efe8"
					},
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"email": {
						"type": "string",
						"example": "johndoe@gmail.com"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email"
				]
			}
		}
	}
}